# 06-job-finder/swagger.yaml
openapi: 3.0.0
info:
  title: Job Sekker
  contact: {}
  version: '1.0'
servers:
- url: https://job-finder-5h1s.onrender.com/api/v1
  variables: {}
paths:
...
  /jobs:
    get:
      tags:
      - Jobs
      summary: Get All Jobs
      operationId: GetAllJobs
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Jobs
      summary: Create Job
      operationId: CreateJob
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateJobRequest'
              - example:
                  company: samsung
                  position: free
            example:
              company: samsung
              position: free
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    CreateJobRequest:
      title: CreateJobRequest
      required:
      - company
      - position
      type: object
      properties:
        company:
          type: string
        position:
          type: string
      example:
        company: samsung
        position: free
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...